Statemint integration tests to cover USDT bidirectional reserve transfer Pendulum <-> Statemint (#196)
* add statemint-runtime to integration tests

* add Statemint to RelayNet as system chain 1000

* add statemine_transfer_asset_to_pendulum unit tests

* check BOB balance in statemine_transfer_asset_to_pendulum test

* add statemine_transfer_asset_to_statemint

* rename statemine to statemint. because this is tests for polkadot net

* add statemint_transfer_incorrect_asset_to_pendulum_fails

* move pendulum and statemint configuration from lib.rs to setup.rs

* move test from lib.rs to test.rs

* move polkadot net configuration to polkadot_test_net

* split tests, setup and polkadot_test_net to different files

* remove unused import in all files

* rename to pendulum-runtime-integration-tests

* Revert "rename to pendulum-runtime-integration-tests"

This reverts commit 2f18be69312712afbe64875988c7674429ca12ca.
